+++ packages/blocks/core/js/wordpress/navigation-submenu/index.js
@@ -11,6 +11,7 @@ import { __ } from '@wordpress/i18n';
 import {
 	SharedBlockExtension,
 	generalBlockStates,
+	generalInnerBlockStates,
 	sharedBlockStates,
 } from '@blockera/editor';
 import { Icon } from '@blockera/icons';
@@ -32,6 +33,18 @@ export const NavigationSubmenu: BlockType = {
 			icon: <Icon icon="link" library="wp" iconSize="20" />,
 			settings: {
 				force: true,
+			},
+			availableBlockStates: {
+				...generalBlockStates,
+				focus: {
+					...generalBlockStates.focus,
+					force: true,
+				},
+				active: {
+					...sharedBlockStates.active,
+					force: true,
+				},
+				visited: sharedBlockStates.visited,
 			},
 		},
 		'elements/submenu-icon': {
@@ -75,88 +88,28 @@ export const NavigationSubmenu: BlockType = {
 			settings: {
 				force: true,
 			},
+			availableBlockStates: {
+				...generalInnerBlockStates,
+				'current-menu-item': {
+					...sharedBlockStates['current-menu-item'],
+					force: true,
+				},
+				'current-menu-parent': sharedBlockStates['current-menu-parent'],
+				'current-menu-ancestor':
+					sharedBlockStates['current-menu-ancestor'],
+				active: sharedBlockStates.active,
+				visited: sharedBlockStates.visited,
+			},
 		},
 	},
 	availableBlockStates: {
 		...generalBlockStates,
 		'current-menu-item': {
-			type: 'current-menu-item',
-			label: __('Current Page', 'blockera'),
-			category: 'special',
-			tooltip: (
-				<>
-					<h5>{__('Is Current Page?', 'blockera')}</h5>
-					<p>
-						{__(
-							'Activates if the block URL is the same as the current page.',
-							'blockera'
-						)}
-					</p>
-					<code style={{ margin: '5px 0' }}>
-						<span style={{ opacity: '0.7' }}>.block</span>
-						.current-menu-item
-					</code>
-				</>
-			),
-			breakpoints: {},
-			priority: 10,
+			...sharedBlockStates['current-menu-item'],
 			force: true,
-			settings: {
-				color: 'var(--blockera-controls-states-color)',
-			},
 		},
-		'current-menu-parent': {
-			type: 'current-menu-parent',
-			label: __('Current Page Parent', 'blockera'),
-			category: 'special',
-			tooltip: (
-				<>
-					<h5>{__('Is Current Page Parent?', 'blockera')}</h5>
-					<p>
-						{__(
-							'Indicates that one of this item’s direct descendants is the current page.',
-							'blockera'
-						)}
-					</p>
-					<code style={{ margin: '5px 0' }}>
-						<span style={{ opacity: '0.7' }}>.block</span>
-						.current-menu-parent
-					</code>
-				</>
-			),
-			breakpoints: {},
-			priority: 10,
-			force: false,
-			settings: {
-				color: 'var(--blockera-controls-states-color)',
-			},
-		},
-		'current-menu-ancestor': {
-			type: 'current-menu-ancestor',
-			label: __('Current Page Ancestor', 'blockera'),
-			category: 'special',
-			tooltip: (
-				<>
-					<h5>{__('Is Current Page Ancestor?', 'blockera')}</h5>
-					<p>
-						{__(
-							'Indicates that one of this item’s descendants is the current page.',
-							'blockera'
-						)}
-					</p>
-					<code style={{ margin: '5px 0' }}>
-						<span style={{ opacity: '0.7' }}>.block</span>
-						.current-menu-ancestor
-					</code>
-				</>
-			),
-			breakpoints: {},
-			priority: 10,
-			force: false,
-			settings: {
-				color: 'var(--blockera-controls-states-color)',
-			},
-		},
+		'current-menu-parent': sharedBlockStates['current-menu-parent'],
+		'current-menu-ancestor': sharedBlockStates['current-menu-ancestor'],
 		active: sharedBlockStates.active,
 		visited: sharedBlockStates.visited,
 	},