+++ packages/blocks/core/js/wordpress/post-comments-form/test/functionality.blocks-2.e2e.cy.js
@@ -25,6 +25,7 @@ describe('Post Comments Form Block', () => {
 
 		// Block supported is active
 		cy.get('.blockera-extension-block-card').should('be.visible');
+
 		//
 		// 1. Edit Block
 		//