+++ packages/blocks/core/js/wordpress/query-pagination-numbers/index.js
@@ -12,6 +12,7 @@ import {
 	SharedBlockExtension,
 	generalBlockStates,
 	sharedBlockStates,
+	generalInnerBlockStates,
 } from '@blockera/editor';
 import { Icon } from '@blockera/icons';
 
@@ -31,6 +32,18 @@ export const QueryPaginationNumbers: BlockType = {
 			icon: <Icon icon="block-pagination-numbers" size="20" />,
 			settings: {
 				force: true,
+			},
+			availableBlockStates: {
+				...generalInnerBlockStates,
+				focus: {
+					...generalInnerBlockStates.focus,
+					force: true,
+				},
+				active: {
+					...sharedBlockStates.active,
+					force: true,
+				},
+				visited: sharedBlockStates.visited,
 			},
 		},
 		'elements/current': {