+++ packages/blocks/core/js/wordpress/query-pagination-previous/test/functionality.blocks-2.e2e.cy.js
@@ -16,6 +16,9 @@ import { testContent } from './test-content';
 
 describe('Query Pagination Previous Block', () => {
 	beforeEach(() => {
+		// Generate 20 posts to make sure the pagination is visible
+		cy.wpCli('wp post generate --count=20 --post_type=post');
+
 		createPost();
 	});