+++ packages/controls/js/libs/search-control/index.js
@@ -74,6 +74,7 @@ export default function SearchControl({
 				onChange={setValue}
 				placeholder={placeholder}
 				className={controlClassNames('search', className)}
+				__nextHasNoMarginBottom={true}
 				{...props}
 			/>
 		</BaseControl>