+++ packages/editor/js/components/editor-advanced-label/test/repeater.advanced-label-3.e2e.cy.js
@@ -341,7 +341,7 @@ describe('Repeater Control label testing (Image & Gradient)', () => {
 
 			// Alias
 			cy.getByAriaLabel('Type').as('type');
-			cy.getByAriaLabel('Angel').as('angel');
+			cy.getByAriaLabel('Angle').as('angle');
 			cy.getByAriaLabel('Effect').as('effect');
 
 			// Assert label in normal state
@@ -354,22 +354,22 @@ describe('Repeater Control label testing (Image & Gradient)', () => {
 		// Assert Type label
 		cy.get('@type').should('have.class', 'changed-in-normal-state');
 
-		// Assert angel label before set value
-		cy.get('@angel').should('not.have.class', 'changed-in-normal-state');
+		// Assert angle label before set value
+		cy.get('@angle').should('not.have.class', 'changed-in-normal-state');
 
 		// Set value
 		cy.getByAriaLabel('Rotate Anti-clockwise').click();
 
-		// Assert Angel label
-		cy.get('@angel').should('have.class', 'changed-in-secondary-state');
+		// Assert Angle label
+		cy.get('@angle').should('have.class', 'changed-in-secondary-state');
 
-		// Assert angel label before set value
+		// Assert effect label before set value
 		cy.get('@effect').should('not.have.class', 'changed-in-normal-state');
 
 		// Set value
 		cy.getByAriaLabel('Parallax').click();
 
-		// Assert Angel label
+		// Assert effect label
 		cy.get('@effect').should('have.class', 'changed-in-secondary-state');
 
 		setBlockState('Normal');
@@ -378,8 +378,8 @@ describe('Repeater Control label testing (Image & Gradient)', () => {
 		// Assert Type label
 		cy.get('@type').should('have.class', 'changed-in-normal-state');
 
-		// Assert Angel label
-		cy.get('@angel').should('have.class', 'changed-in-other-state');
+		// Assert angle label
+		cy.get('@angle').should('have.class', 'changed-in-other-state');
 
 		// Assert Effect label
 		cy.get('@effect').should('have.class', 'changed-in-other-state');
@@ -399,8 +399,8 @@ describe('Repeater Control label testing (Image & Gradient)', () => {
 		// Assert Type label
 		cy.get('@type').should('have.class', 'changed-in-normal-state');
 
-		// Assert Angel label
-		cy.get('@angel').should('have.class', 'changed-in-other-state');
+		// Assert angle label
+		cy.get('@angle').should('have.class', 'changed-in-other-state');
 
 		// Assert Effect label
 		cy.getByAriaLabel('Effect').should(
@@ -420,8 +420,8 @@ describe('Repeater Control label testing (Image & Gradient)', () => {
 			'changed-in-normal-state'
 		);
 
-		// Assert Angel label
-		cy.get('@angel').should('have.class', 'changed-in-other-state');
+		// Assert angle label
+		cy.get('@angle').should('have.class', 'changed-in-other-state');
 	});
 
 	describe('reset action testing...', () => {