+++ packages/editor/js/extensions/libs/block-composite/components/categories.js
@@ -3,7 +3,7 @@
 /**
  * External dependencies
  */
-import { __ } from '@wordpress/i18n';
+import { __, sprintf } from '@wordpress/i18n';
 import { select } from '@wordpress/data';
 import type { MixedElement } from 'react';
 import { useState } from '@wordpress/element';
@@ -222,7 +222,7 @@ export const Categories = ({
 	};
 
 	return (
-		<Flex direction={'column'} gap={'15px'}>
+		<Flex direction={'column'} gap={'20px'}>
 			<ControlContextProvider
 				value={{
 					name: 'search-block-' + clientId,
@@ -436,7 +436,11 @@ export const Categories = ({
 					<CategorizedItems
 						{...categorizedItemsProps}
 						itemType={'inner-block'}
-						title={title}
+						title={sprintf(
+							// translators: %s is the category title
+							__('%s Blocks', 'blockera'),
+							title
+						)}
 						limited={true}
 						items={blocks}
 						category={slug}