+++ packages/editor/js/extensions/libs/block-composite/components/categorized-items.js
@@ -128,7 +128,7 @@ export const CategorizedItems = ({
 
 			<Grid
 				gridTemplateColumns={'repeat(2, 1fr)'}
-				gap={'5px'}
+				gap={'8px'}
 				className={`blockera-features-types blockera-feature-${category}-wrapper`}
 			>
 				{validItems.map(
@@ -283,7 +283,11 @@ export const CategorizedItems = ({
 									}
 									tabIndex="0"
 									role="button"
-									data-test={name}
+									data-test={
+										itemType === 'state'
+											? 'states/' + id
+											: name
+									}
 									className={classNames(
 										'blockera-feature-type',
 										'is-item'